backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0.74
Pillow>=9.5.0
numpy<2.0.0
scikit-learn>=1.3.0
Werkzeug>=2.3.0
python-multipart>=0.0.6
